完成接口装饰器化改造，添加Loading组件，完成顶栏历史记录获取，图片上传的前端部分

+++ src/views/Create/components/CreateVideo.vue
@@ -85,6 +85,7 @@ export default defineComponent({
     }
 
     const deleteFile = async () => {
+      if (fileHash.value === '') return
       try {
         await VideoApi.deleteVideoFile(fileHash.value)
         fileHash.value = ''
@@ -127,7 +128,6 @@ export default defineComponent({
 
     onBeforeRouteLeave(async (to) => {
       try {
-        console.log(to.fullPath, '/create')
         if (to.fullPath !== '/create') await cancel()
       } catch {
         return false
